refactor /sale & /deposit. Decimals.

+++ src/models/database_models.py
@@ -66,14 +66,14 @@ class PayOrder(Base):
 
     # Input payment details
     source_currency_id = Column(String, nullable=True)
-    source_amount = Column(Float, nullable=True) 
+    source_amount = Column(BigInteger, nullable=True) 
     source_value_usd = Column(Float, nullable=True)
     source_transaction_hash = Column(String, nullable=True)
     source_deposit_address = Column(String, nullable=True)
     
     # Output payment details
     destination_currency_id = Column(String, nullable=True)
-    destination_amount = Column(Float, nullable=True)
+    destination_amount = Column(BigInteger, nullable=True)
     destination_address = Column(String, nullable=True)
     destination_value_usd = Column(Float, nullable=True)
     destination_transaction_hash = Column(String, nullable=True)
